requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
llama-index>=0.10.15
llama-parse>=0.3.8
llama-index-vector-stores-qdrant>=0.1.3
llama-index-llms-openai>=0.1.6
llama-index-embeddings-gemini>=0.1.6
llama-index-tools-tavily-research>=0.1.0
llama-index-llms-gemini>=0.1.6
qdrant-client>=1.7.3
llama-index-retrievers-bm25>=0.7.0
python-dotenv>=1.0.1
pydantic>=2.6.1
nest_asyncio>=1.6.0
loguru>=0.7.2
prometheus-fastapi-instrumentator>=7.0.0
arize-phoenix>=4.0.0
openinference-instrumentation-llama-index>=2.0.0
redis>=5.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
PyJWT[crypto]>=2.8.0
langchain-ollama>=0.2.0
